Make HBHandler<T,S> struct instead of trait Make HeartbeatHandler a generic struct instead of a struct with a trait. Reduces redundant code
